fix: Type errors and test corrections
- PipelineRun::duration() cast to int for Carbon compatibility
- ResourceAlert::getCooldownRemainingMinutesAttribute() cast to int
- TenantDeployment tests use Tenant::factory() with database field
- Policy tests corrected to expect proper ownership restrictions
- Server/Project status tests create resources with correct user_id

+++ app/Models/ResourceAlert.php
@@ -137,7 +137,7 @@ class ResourceAlert extends Model
             return null;
         }
 
-        return now()->diffInMinutes($this->last_triggered_at->addMinutes($this->cooldown_minutes));
+        return (int) now()->diffInMinutes($this->last_triggered_at->addMinutes($this->cooldown_minutes));
     }
 
     /**
